# SERVER/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/genere": {
      "get": {
        "description": "Questa route serve per richiedere la lista dei generi utilizzando le api di spotyfi. Ritorna 200 e la lista di generi.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/elimina": {
      "delete": {
        "description": "Questa route serve per eliminare un profilo. In caso l'eliminazione venga eseguita con sucesso rotorna 200, nel caso si sia verificato un errore in fase di risposta o eliminazione manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/modplaylist1": {
      "post": {
        "description": "Dato un token (attivo) di un profilo restutuisce le playlist create e ancora attive di quel profilo (200), nel caso si sia verificato un errore in fase di risposta o ricerca manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/modplaylist2": {
      "post": {
        "description": "Dato un token (attivo) di un profilo e un nome di una playlist creata da quel profilo e ancora attiva trova e ritorna le canzoni della playlist cercata (200), nel caso si sia verificato un errore in fase di ricerca o risposta manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "playlist": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }

      }
    },
    "/modplaylist3": {
      "post": {
        "description": "Dato un token (attivo) di un profilo trova e ritorna le playlist create e non da quel profilo ancora attive (200), nel caso si sia verificato un errore in fase di ricerca o risposta manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
   "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/modplaylist4": {
      "post": {
        "description": "Dato un token (attivo) di un profilo e un nome di una playlist creata o non da quel profilo e ancora attiva. Trova e ritorna le canzoni della playlist cercata (200), nel caso si sia verificato un errore in fase di ricerca o risposta manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "playlist": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/eliminaPlaylist": {
      "delete": {
        "description": "Dato un token (attivo) di un profilo e un nome di una playlist elimina la playlist di quel profilo (200), nel caso si sia verificato un errore in fase di eliminazione o risposta manda uno status 500. Tutti gli altri utenti che si sono salvati la playlist non la vedranno più.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/togliPlaylist": {
      "delete": {
        "description": "Dato un token (attivo) di un profilo e un nome di una playlist toglie la playlist di quel profilo (200), nel caso si sia verificato un errore in fase di eliminazione o risposta manda uno status 500. La playlist rimarrà attiva negli altri profili in cui è stata salvata.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/modplaylist5": {
      "post": {
        "description": "Dato un token (attivo) di un profilo trova e ritorna le playlist non create e non aggiunte dall'utente (200), nel caso si sia verificato un errore in fase di ricerca o risposta manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/modplaylist6": {
      "post": {
        "description": "Dato un token (attivo) di un profilo e un nome di una playlist non creata da quel profilo e ancora attiva trova e ritorna le canzoni della playlist cercata (200), nel caso si sia verificato un errore in fase di ricerca o risposta manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "emailpass": {
                  "example": "any"
                },
                "playlist": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/register": {
      "put": {
        "description": "Registra un nuovo utrente nella piattaforma (200), nel caso ci siano problemi in fase di inserimento manda uno status 500, nel casoci siano problemi di controllo dei dati inseriti manda uno status 400.",
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "Effettua il login di un utrente nella piattaforma (200), nel caso ci siano problemi in fase di connessione manda uno status 500, nel casoci siano problemi di controllo dei dati inseriti manda uno status 400. Inoltre crea il token di accesso, lo salva nel server assieme al timestamp e lo manda all'utente (per chiamate future).",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "description": "Effettua il logout di un utrente dalla piattaforma (200) rimuovendo il token dalla lista dei token presenti.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "unauthorized"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/mod": {
      "post": {
        "description": "Dato un token (attivo) di un profilo restituisce i dati di tale profilo (200),nel caso ci siano problemi in fase di connessione manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/modPass": {
      "put": {
        "description": "Dato un token (attivo) di un profilo , la nuova password e la vecchia password imposta la password a quella nuova (200), nel caso ci siano problemi in fase di connessione manda uno status 500, nel casoci siano problemi di controllo dei dati inseriti manda uno status 400.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/modData": {
      "put": {
        "description": "Dato un token (attivo) di un profilo e un json di campi da modificare aggiorna i dati del profilo (200), nel caso ci siano problemi in fase di connessione manda uno status 500, nel casoci siano problemi di controllo dei dati inseriti manda uno status 400.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/ADDplaylist": {
      "put": {
        "description": "Dato un token (attivo) di un profilo, una mail del creante di una playlist e il nome della playlist al quale mi voglio aggiungere, aggiungo al mio profilo la playlist (200), nel caso ci siano problemi in fase di connessione o aggiunta manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "emailpass": {
                  "example": "any"
                },
                "playlist": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/cerca": {
      "post": {
        "description": "Dato un token (attivo) di un profilo e un dato da cercare (artista, nome canzone, album) cercca le canzoni più simili a quel caso e le restituisce(200), nel caso ci siano problemi in fase di connessione o ricerca manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cercato": {
                  "example": "any"
                },
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/artisti": {
      "post": {
        "description": "Dato un artista da cercare, lo cerca e le restituisce(200), nel caso ci siano problemi in fase di connessione o ricerca manda uno status 500.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cercato": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/salva": {
      "put": {
        "description": "Dato un token (attivo) di un profilo e un json di una playlist, la salva nel profilo(200). Nel caso ci siano problemi in fase di connessione o savltaggio manda uno status 500, nel caso ci siano problemi di controllo dei dati inseriti manda uno status 400.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/salvaMod": {
      "put": {
        "description": "Dato un token (attivo) di un profilo e un json di una playlist, salva le modifiche nel profilo(200). Nel caso ci siano problemi in fase di connessione o salvataggio manda uno status 500, nel caso ci siano problemi di controllo dei dati inseriti manda uno status 400.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "unauthorized"
          },
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/forgot": {
      "post": {
        "description": "Riceve una richiesta di password forgot",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "bed request"
          },
          "200": {
            "description": "ok"
          }
        }
      }
    }
  }
}